Fixed rendering & stochastic simulation bugs
The PySCeS file was not being generated correctly for the stochastic
simulator.

Pyplot would have some strange behavior when plotting depending on
whether or not the user was showing interactively or saving to an
image. Now the same configuration should work for users over ssh and
users with an X-Server.